requirements.txt
flask
flask-sqlalchemy
flask-wtf
wtforms
email_validator
flask_bcrypt
flask_login
python-dotenv